Today
Read last days one and then ne updates: Worked on Newbackings syem need to make levle button work and test eqquip button right now when eqqip is presed it will be set to NBsprit from UpadeBackfgound in HM2 and Newbacking Equp folder. + Need to fished sytem and add it to all other aspects and zones and levles and stuff and still nned to make all the thing clear and exated backgounds. + nned to add in all new bakgounds.